Got the GUI working display the game in progress!
I had to make gameEngine run in a new thread.  Phill you might want to change the evaluateFitnessContest method so it uses the new method for running the simulation rather than the old one, so it can be deleted.

There are some issues with speed when it comes to zooming and panning.  I'll try and think of ways of improving performance, but the only thing I can think of is by somehow converting the unmoving elements into one big image, and then drawing that.