Done a lot of work towards optimising efficiency.  In he end, after trying alternatives, I decided to have one size for the background tiles and buffering this.  This meant that I could get rid of all the rock tile shade mapping code (I also did the tinting of these tiles programatically now rather than having a different image for each shade).  I could also get rid of all the images for each scale size of the background now.
Basically, the program now takes a little while to load, but the game runs much faster while it is actually displaying.

In hindsight I should have made the images smaller, so it wouldn't have produced such massively large off screen buffered backgrounds.  But this would require so much work to change now because a lot of the zooming and scaling has all been hard coded in.

Next, I think I will get rid of the medium size ant scales and just use the large scale (or at least try that).